Refactoring by feedback
