+++ stock-system/src/main/java/saffchen/command/ReceiverDB.java
@@ -54,33 +54,4 @@ public class ReceiverDB {
         }
     }
 
-    public void createXLS() throws GeneralSecurityException, IOException {
-        System.out.println("Укажите название файла");
-        System.out.println("Имя файла не должно содержать символы: /, |, ?, *, :, <>");
-        String name = new Scanner(System.in).nextLine();
-        while (name.contains("?") || name.contains("/") || name.contains(":") || name.contains("*") ||
-                name.contains("\"") || name.contains("<") || name.contains(">") || name.contains("|") || name.contains("\\")) {
-            System.out.println("Имя файла содержит запрещенные символы");
-            System.out.println("Введите имя");
-            name = new Scanner(System.in).nextLine();
-        }
-        String path = System.getProperty("user.home") + "\\Desktop\\" + name + ".xls";
-        Sheets sheetsService = getSheetsService();
-        String range = "Sheet1!A1:Z1000";
-        ValueRange response = sheetsService.spreadsheets().values().get(SPEADSHEET_ID, range).execute();
-        List<List<Object>> values = response.getValues();
-        Workbook book = new HSSFWorkbook();
-        Sheet sheetONE = book.createSheet("Sheet1");
-        List<Row> rows = new ArrayList<>();
-        int i = 0;
-        for (List<Object> rower : values) {
-            rows.add(sheetONE.createRow(i));
-            for (int j = 0; j < rower.size(); j++) {
-                rows.get(i).createCell(j).setCellValue(rower.get(j).toString());
-            }
-            i++;
-        }
-        book.write(new FileOutputStream(path));
-        book.close();
-    }
 }
